requirements.txt
openai==1.99.1
pandas==2.3.1
numpy==2.2.6
matplotlib==3.10.3
seaborn==0.13.2
python-dateutil==2.9.0.post0
openpyxl==3.1.5
python-dotenv==1.1.1
pydantic==2.11.7